+++ example/ios/Dummy.swift
@@ -1,0 +1,6 @@
+//
+//  Dummy.swift
+//  example
+//
+
+import Foundation