+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Example project in the `example` directory.
+
+### Changed
+
+- Color properties (`toolbarColor`, `preferredBarTintColor` and `preferredControlTintColor`) now
+  accept more color formats. Please refer to TinyColor's [Accepted String Input](https://github.com/bgrins/TinyColor#accepted-string-input) to see all available formats.
+
 ## [1.2.7] - 2019-01-29
 
 ### Added